+++ clx_budget_analysis_report/wizard/budget_report.py
@@ -131,6 +131,8 @@ class BudgetReportWizard(models.TransientModel):
                 for sub_line in starting_line:
                     if sub_line.end_date and subscription_line.start_date >= sub_line.end_date:
                         starting_line -= sub_line
+                    elif sub_line.end_date and sub_line.end_date < end_date_x:
+                        starting_line -= sub_line
                 s_sum = 0.0
                 for s_line in starting_line:
                     if current_month_start_date.day > 15 or end_date_x.day < 15: